mimi_claw_pi/soul.py
import os
import json
import google.generativeai as genai
from datetime import datetime

# Setup Gemini
# Replace with your actual key or use an env variable
# genai.configure(api_key="YOUR_GEMINI_API_KEY")
# model = genai.GenerativeModel('gemini-2.0-flash-exp')

class MimiSoul:
    def __init__(self):
        self.memory_path = "mimi_claw_pi/memory/"
        self.ui_path = "mimi_claw_pi/ui/"
        
    def load_memory(self, filename):
        path = os.path.join(self.memory_path, filename)
        if os.path.exists(path):
            with open(path, 'r') as f:
                return f.read()
        return ""

    def save_memory(self, filename, content):
        with open(os.path.join(self.memory_path, filename), 'w') as f:
            f.write(content)

    def trigger_ui_alert(self, message, action="Got it"):
        alert_data = {"message": message, "action": action, "timestamp": str(datetime.now())}
        with open(os.path.join(self.ui_path, "latest_alert.json"), 'w') as f:
            json.dump(alert_data, f)

    def think(self, user_input):
        # 1. Load context
        soul = self.load_memory("SOUL.md")
        user = self.load_memory("USER.md")
        
        # 2. Build Prompt (Mocking Gemini call for now)
        prompt = f"System: {soul}\nUser Info: {user}\nInput: {user_input}"
        
        # 3. Process (Simulated)
        response = f"Simulated Gemini Response to: {user_input}"
        
        # 4. Handle UI interaction if needed
        if "alert" in response.lower() or "question" in response.lower():
            self.trigger_ui_alert(response, action="Reply via Telegram")
            
        return response

if __name__ == "__main__":
    soul = MimiSoul()
    # Example usage:
    # soul.think("Remind me to water the plants tomorrow.")
