mimi_claw_pi/requirements.txt
google-generativeai
kivy
pyyaml
requests
python-dotenv
